Take out unneeded code and make code much safer
Take out the wn.mainloop(), as it's not doing anything, hiding under your main loop and never being executed till the program quits, and make the code safer with the " if __name__ == '__main__': " thing, which if it runs as a module (no idea why you would even do that anyways) the program might break stuff. Also loops totally can use "_" as a counter, but "_" can also cause special things in python that you may not want to happen. Also fixed a bit of spacing and made the window title a bit better. Finally, I also made it more efficient by making the rim lines and hands into functions, instead of taking up about 6 times the space with very similar code.